Mit ausnahmebehandlung

+++ parent-project-skeleton/client/src/main/java/at/jku/Controller.java
@@ -59,6 +59,8 @@ public class Controller implements Initializable {
     @FXML
     private Label peopleInRoomLabel;
     @FXML
+    private Label fehlermeldungLabel;
+    @FXML
     private Label createdLight;
     @FXML
     private Label roomUpdated;
@@ -243,6 +245,17 @@ public class Controller implements Initializable {
             createdLight.setText("Bitte Raum auswählen");
             return;
         }
+
+            if(lightId.getText().length() == 0){
+                createdLight.setText("Fill in all required values!");
+                return;
+            }
+
+            if(lightName.getText().length() == 0){
+                createdLight.setText("Fill in all required values!");
+                return;
+            }
+
         currentRoom = getCompleteRoom(comboBox1.getValue().getRoom_id());
         String lightName = getNameOfLight();
         String lightId = getLightId();
@@ -258,6 +271,15 @@ public class Controller implements Initializable {
             return;
         }
 
+        if(roomSize.getText().length() == 0){
+            createdLight.setText("Fill in all required values!");
+            return;
+        }
+        if(roomSize.getText().getClass().getSimpleName().equals("String")){
+            createdLight.setText("Fill in a number!");
+            return;
+        }
+
         currentRoom = getCompleteRoom(comboBox1.getValue().getRoom_id());
 
         client.updateRoom(currentRoom.getRoom_id(), getRoomSize(), globalMeasurementUnit);
@@ -270,6 +292,16 @@ public class Controller implements Initializable {
 
         if(comboBox1.getValue() == null) {
             createdLight.setText("Bitte Raum auswählen");
+            return;
+        }
+
+        if(fanName.getText().length() == 0){
+            createdLight.setText("Fill in all required values!");
+            return;
+        }
+
+        if(fanId1.getText().length() == 0){
+            createdLight.setText("Fill in all required values!");
             return;
         }
 
@@ -291,6 +323,16 @@ public class Controller implements Initializable {
             return;
         }
 
+        if(doorName.getText().length() == 0){
+            createdLight.setText("Fill in all required values!");
+            return;
+        }
+
+        if(doorId.getText().length() == 0){
+            createdLight.setText("Fill in all required values!");
+            return;
+        }
+
         currentRoom = getCompleteRoom(comboBox1.getValue().getRoom_id());
 
         String doorName = getNameOfDoor();
@@ -306,6 +348,16 @@ public class Controller implements Initializable {
 
         if(comboBox1.getValue() == null) {
             createdLight.setText("Bitte Raum auswählen");
+            return;
+        }
+
+        if(windowId.getText().length() == 0){
+            createdLight.setText("Fill in all required values!");
+            return;
+        }
+
+        if(windowName.getText().length() == 0){
+            createdLight.setText("Fill in all required values!");
             return;
         }
 
@@ -510,46 +562,56 @@ public class Controller implements Initializable {
     @FXML
     public void deleteRoom() throws IOException {
 
-        Room_Object room_object = roomTableView.getSelectionModel().getSelectedItem();
+        if(roomTableView.getSelectionModel().getSelectedItem() == null)
+            fehlermeldungLabel.setText("Select a room!");
+        else {
 
-        client.deleteRoom(room_object.getRoom_id());
+            Room_Object room_object = roomTableView.getSelectionModel().getSelectedItem();
 
-        rooms.clear();
+            client.deleteRoom(room_object.getRoom_id());
 
-        roomTableView.setItems(getRooms());
+            rooms.clear();
+
+            roomTableView.setItems(getRooms());
+        }
     }
 
     @FXML
     public void deleteComponent () {
 
-        Component component = detailTableView.getSelectionModel().getSelectedItem();
+        if(detailTableView.getSelectionModel().getSelectedItem() == null)
+            fehlermeldungLabel.setText("Select a component!");
+        else {
 
-        client.deleteRoomLight(component.getRoom_id(), component.getId());
-        client.deleteVent(component.getRoom_id(), component.getId());
-        client.deleteDoor(component.getRoom_id(), component.getId());
-        client.deleteWindow(component.getRoom_id(),component.getId());
+            Component component = detailTableView.getSelectionModel().getSelectedItem();
 
-        details.clear();
+            client.deleteRoomLight(component.getRoom_id(), component.getId());
+            client.deleteVent(component.getRoom_id(), component.getId());
+            client.deleteDoor(component.getRoom_id(), component.getId());
+            client.deleteWindow(component.getRoom_id(), component.getId());
 
-        showRoom();
+            details.clear();
+
+            showRoom();
+        }
     }
     @FXML
     public void changeStatus () {
 
-        Component component = detailTableView.getSelectionModel().getSelectedItem();
+        if(detailTableView.getSelectionModel().getSelectedItem() == null)
+            fehlermeldungLabel.setText("Select a component!");
+        else {
 
+            Component component = detailTableView.getSelectionModel().getSelectedItem();
 
-        component.changeStatus();
+            component.changeStatus();
 
-        //boolean status = component.getStatus();
+            System.out.println(component.getStatus()); //zum testen
 
-        //component.setStatus( (!status) );
+            details.clear();
 
-        System.out.println(component.getStatus()); //zum testen
-
-        details.clear();
-
-        showRoom();
+            showRoom();
+        }
 
     }
 
@@ -607,10 +669,15 @@ public class Controller implements Initializable {
     @FXML
     public void createNewRoom() throws IOException {
 
-       Room_Object r = client.addRoom(getRoomId(), getRoomSize(), globalMeasurementUnit);
-       //client.addPeopleRoom(getRoomId(), 0);
+        if(roomSize.getText().length() == 0 || roomId.getText().length() == 0)
+            createdRoom.setText("Fill in all required values!");
+        else {
 
-       createdRoom.setText(getRoomId() + " wurde erstellt.");
+            Room_Object r = client.addRoom(getRoomId(), getRoomSize(), globalMeasurementUnit);
+            //client.addPeopleRoom(getRoomId(), 0);
+
+            createdRoom.setText(getRoomId() + " wurde erstellt.");
+        }
     }
